現在のnoteListクラスを削除
