--- 04.class/lib/note_list.js
@@ -1,28 +1,0 @@
-import Enquirer from "enquirer";
-
-export class NoteList {
-  constructor(notes) {
-    this.notes = notes;
-  }
-
-  seeAllTitles() {
-    this.notes.forEach((note) => console.log(note.title));
-  }
-
-  async selectNote(messageText) {
-    const question = {
-      type: "select",
-      name: "toSeeNote",
-      message: messageText,
-      choices: this.notes.map((note) => ({
-        name: note.title,
-        value: note,
-      })),
-      result() {
-        return this.focused.value;
-      },
-    };
-    const answer = await Enquirer.prompt(question);
-    return answer.toSeeNote;
-  }
-}
